created notifications tab
